refactor(orchestration): keep mod.rs export-focused per module convention
Move the wake-graph runtime seam, builder, run entrypoint and topology out of
graph/mod.rs into a graph/build.rs sibling, and the steering task-local helpers
out of reasoning_agent/mod.rs into reasoning_agent/steering.rs. Both mod.rs
files are now export-only (mod/pub mod + pub use), matching the project's
canonical module shape. Also refresh the cycle_id field doc to note the
counterpart_agent_id scoping.

Addresses @coderabbitai on graph/mod.rs:417 and reasoning_agent/mod.rs:36.
